Callback can take more than five seconds
Under :sync t, we waited five seconds for semaphore to flip, but an
arbitrary :success or :complete callback could take longer than that.
This resulted in "semaphore never called".

We use the fact that process sentinel is temporarily set to nil when
running to avoid debiting max iterations while sentinel is taking its
sweet time.